First entry
This is what I've done so far with the site. I just found out about mobile resizing so I made a new css file to make it responsive outside of the media queries.
Still some work to be done on some more pages before I can focus on the JS and mobile navbar